🔧[refactor]: Apple 로그인 DI 통합 및 UI 개선
- Apple 인증 DI 의존성 주입 완료
- SignInWithAppleButton UI 레이어링 수정
- AppleLoginManager를 DefaultAppleAuthRequestImpl로 통합
