+++ Projects/Domain/DomainInterface/Sources/OAuth/AppleAuthRequestInterface.swift
@@ -14,22 +14,23 @@ public protocol AppleAuthRequestInterface: Sendable {
 }
 
 ///// OAuth Repository의 DependencyKey 구조체
-//public struct  AppleAuthRequestDependency: DependencyKey {
-//  public static var liveValue: AppleAuthRequestInterface {
-//    UnifiedDI.resolve(AppleAuthRequestInterface.self) ?? DefaultAppleAuthRequest()
-//  }
-//
-//  public static var testValue: AppleAuthRequestInterface {
-//    UnifiedDI.resolve(AppleAuthRequestInterface.self) ?? DefaultAppleAuthRequest()
-//  }
-//
-//  public static var previewValue: AppleAuthRequestInterface = liveValue
-//}
-//
-///// DependencyValues extension으로 간편한 접근 제공
-//public extension DependencyValues {
-//  var appleManger: AppleAuthRequestInterface {
-//    get { self[AppleAuthRequestDependency.self] }
-//    set { self[AppleAuthRequestDependency.self] = newValue }
-//  }
-//}
+public struct  AppleAuthRequestDependency: DependencyKey {
+  public static var liveValue: AppleAuthRequestInterface {
+    UnifiedDI.resolve(AppleAuthRequestInterface.self) ?? DefaultAppleAuthRequestImpl()
+  }
+
+  public static var testValue: AppleAuthRequestInterface {
+    UnifiedDI.resolve(AppleAuthRequestInterface.self) ?? DefaultAppleAuthRequestImpl()
+  }
+
+  public static var previewValue: AppleAuthRequestInterface = liveValue
+}
+
+/// DependencyValues extension으로 간편한 접근 제공
+public extension DependencyValues {
+  var appleManger: AppleAuthRequestInterface {
+    get { self[AppleAuthRequestDependency.self] }
+    set { self[AppleAuthRequestDependency.self] = newValue }
+  }
+}
+
