+++ Projects/Presentation/Auth/Sources/Components/SocialCircleButton.swift
@@ -22,17 +22,7 @@ struct SocialCircleButtonView: View {
     switch type {
     case .apple:
       ZStack {
-        Circle()
-          .fill(.black)
-          .frame(width: circleSize, height: circleSize)
-          .shadow(color: .gray40, radius: 5, x: 0, y: 0)
-
-        Image(systemName: type.image)
-          .resizable()
-          .scaledToFit()
-          .frame(width: 18, height: 30)
-          .foregroundColor(.white)
-
+        // 🔥 테스트: SignInWithAppleButton을 맨 앞으로
         SignInWithAppleButton(.signIn) { request in
           store.send(.async(.prepareAppleRequest(request)))
         } onCompletion: { result in
@@ -40,14 +30,21 @@ struct SocialCircleButtonView: View {
         }
         .frame(width: circleSize, height: circleSize)
         .clipShape(Circle())
-        .opacity(0.02)
         .allowsHitTesting(true)
-        .scaleEffect(isPressed ? 0.95 : 1.0)
-        .onLongPressGesture(minimumDuration: 0, maximumDistance: .infinity, pressing: { pressing in
-          withAnimation(.spring(response: 0.3, dampingFraction: 0.6)) {
-            isPressed = pressing
-          }
-        }, perform: {})
+
+        // 배경 원과 아이콘을 뒤에 배치
+        Circle()
+          .fill(.black)
+          .frame(width: circleSize, height: circleSize)
+          .shadow(color: .gray40, radius: 5, x: 0, y: 0)
+          .allowsHitTesting(false)
+
+        Image(systemName: type.image)
+          .resizable()
+          .scaledToFit()
+          .frame(width: 18, height: 30)
+          .foregroundColor(.white)
+          .allowsHitTesting(false)
       }
       .scaleEffect(isPressed ? 0.95 : 1.0)
       .animation(.spring(response: 0.52, dampingFraction: 0.94, blendDuration: 0.14), value: isPressed)
