+++ Projects/Presentation/Auth/Sources/Login/Reducer/Login.swift
@@ -95,6 +95,7 @@ public struct Login {
   @Dependency(\.authUseCase) var authUseCase
   @Dependency(\.signUpUseCase) var signUpUseCase
   @Dependency(\.profileUseCase) var profileUseCase
+  @Dependency(\.appleManger) var appleLoginManger
   @Dependency(\.continuousClock) var clock
   @Dependency(\.mainQueue) var mainQueue
   
@@ -139,7 +140,7 @@ public struct Login {
   ) -> Effect<Action> {
     switch action {
       case .prepareAppleRequest(let request):
-        let nonce = AppleLoginManager().prepare(request)
+        let nonce = appleLoginManger.prepare(request)
         state.nonce = nonce
         return .none
         
@@ -401,77 +402,5 @@ public struct Login {
       case .presentMemberMain:
         return .none
     }
-  }
-}
-
-
-//
-//  AppleLoginManager.swift
-//  Presentation
-//
-//  Created by Wonji Suh  on 10/29/24.
-//
-
-import CryptoKit
-import Foundation
-import AuthenticationServices
-
-public protocol AppleAuthRequestPreparing {
-  func prepare(_ request: ASAuthorizationAppleIDRequest) -> String
-}
-
-public struct AppleLoginManager: AppleAuthRequestPreparing {
-
-  public init() {}
-
-  public func prepare(_ request: ASAuthorizationAppleIDRequest) -> String {
-    let nonce = randomNonceString()
-    request.requestedScopes = [.email, .fullName]
-    request.nonce = sha256(nonce)
-    return nonce
-  }
-
-  public func sha256(_ input: String) -> String {
-    let inputData = Data(input.utf8)
-    let hashedData = SHA256.hash(data: inputData)
-    let hashString = hashedData.compactMap {
-      String(format: "%02x", $0)
-    }.joined()
-
-    return hashString
-  }
-
-  public func randomNonceString(length: Int = 32) -> String {
-    precondition(length > 0)
-    let charset: [Character] =
-      Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-    var result = ""
-    var remainingLength = length
-
-    while remainingLength > 0 {
-      let randoms: [UInt8] = (0 ..< 16).map { _ in
-        var random: UInt8 = 0
-        let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-        if errorCode != errSecSuccess {
-          fatalError(
-            "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-          )
-        }
-        return random
-      }
-
-      randoms.forEach { random in
-        if remainingLength == 0 {
-          return
-        }
-
-        if random < charset.count {
-          result.append(charset[Int(random)])
-          remainingLength -= 1
-        }
-      }
-    }
-
-    return result
   }
 }
